refactor: Generator 파일 패키지 이동
서비스 계층의 패키지 위치로 인하여 빌드 오류 발생
