+++ src/main/java/bridge/BridgeMaker.java
@@ -1,7 +1,5 @@
 package bridge;
 
-import bridge.service.BridgeNumberGenerator;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
