+++ src/main/java/bridge/PlayerSetting.java
@@ -1,12 +1,10 @@
 package bridge;
 
-import bridge.BridgeMaker;
 import bridge.domain.Player;
 import bridge.dto.BridgeCorrectLocationDto;
 import bridge.dto.BridgeDto;
 import bridge.dto.PlayerBridgeSizeDto;
 import bridge.exception.InputException;
-import bridge.service.BridgeRandomNumberGenerator;
 
 import java.util.List;
